requirements.txt
boto3==1.24.74
botocore==1.27.74
Django==3.2.15
django-bootstrap3==22.1
django-filter==22.1
django-tables2==2.4.1
djangorestframework==3.13.1
Jinja2==2.11.2
jmespath==1.0.1
joblib==0.17.0
jsonschema==3.2.0
numpy==1.19.4
oauthlib==3.1.0
Pillow==8.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
ptyprocess==0.6.0
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.6.0

